pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.39",
    "flask-uploads>=0.2.1",
    "python-magic>=0.4.27",
    "twilio>=9.5.0",
    "requests>=2.32.3",
    "pytube>=15.0.0",
    "flask-babel>=4.0.0",
]
